Bug fix
This is a fix/ correction of all warings that was present in the project.

+++ 3DProject/3DProject/Defines.h
@@ -13,8 +13,8 @@
 //My Define Values
 #define PI 3.14159265359f
 #define ROTATION -1.0f
-#define WIDTH 1280.0f
-#define HEIGHT 800.0f
+#define WIDTH (GLint)1280.0f
+#define HEIGHT (GLint)800.0f
 #define STB_IMAGE_IMPLEMENTATION
 
 #endif
